Add Services (#37)
Services are similar to actions, but they are persistent. Use case is
for servers and long run processes which you want to control.

## Example

The following config run the dev app (next, rails, http servers) and
keeps track of the status. The logs are in consistent locations and
users can inspect their running services.

```yaml
services:
  dev:
    command: pnpm run dev --port $PORT
    port: 3000
    domain: myapp.local
```